Provide mechanism to block on API calls.
Issue is that join cluster is currently an asynchronous call. After the call
returns to the client, the client does not know whether the join completed
successfully without polling to query the cluster. However, even if the client
did poll afterwards, there is still the problem that if the join failed then
the client has no way of distinguishing that from join taking a long time to
execute. This patch provides a mechanism to create synchronous calls that block
until the command has finished.

Fixes #1