Revert "Expand decision logic to evict plus partition"
This reverts commit 7d7c2cc241ab5f2ee48d2406a6d5f765ff1a9721.

Issue is that current logic with this change can lead to situations where the
wrong node is chosen to heal during automatic healing. While it is true that we
should be able to expand the logic it is not as simple because one replica set
will have detected a failed replica, but an entirely differnt replica set will
have to be used to pick a node to heal over.