Fix race-condition no-progress partition endpoint
Issue results from the difference between dafs replica set and paxos
replica set. Before this commit ping used dafs replica set details to
decide whether or not to send a ping proposal. This is problematic
because when a new node is added we are not guaranteed that all the
replicated files have been received before the dafs replica set details
is updated. Essentially, this means that a ping may be proposed before a
replica has all the updated state necessary to send a valid proposal.

Fixes #5